modified footer to perfect almost
