+++ src/components/Footer/Footer.module.scss
@@ -1,0 +1,27 @@
+.footer {
+  @apply border-t border-element-color mt-20 flex justify-center items-center;
+
+  .container {
+    @apply grid grid-rows-1 gap-8 pt-8 pb-8 pl-4 md:flex md:justify-between md:pl-8 md:pr-8 max-w-7xl w-full;
+  }
+
+  .links {
+    @apply grid gap-3 md:flex md:items-center md:ml-[5%] md:gap-14;
+
+    .link {
+      @apply text-secondary-primary uppercase font-mont font-extrabold text-xs;
+    }
+  }
+
+  .button {
+    @apply flex items-center justify-center;
+
+    .button_top {
+      @apply flex items-center self-center gap-4;
+
+      .button_title {
+       @apply text-secondary-primary font-mont text-xs font-bold;
+      }
+    }
+  }
+}
